+++ lib/money/application.ex
@@ -3,8 +3,6 @@ defmodule Money.Application do
   alias Money.ExchangeRates
   require Logger
 
-  @start_service_by_default? false
-
   def start(_type, _args) do
     import Supervisor.Spec, warn: false
 
@@ -23,11 +21,10 @@ defmodule Money.Application do
   defp start_exchange_rate_service? do
     maybe_log_deprecation()
 
-    start? = Money.get_env(:auto_start_exchange_rate_service, @start_service_by_default?)
     api_module = ExchangeRates.default_config().api_module
     api_module_present? = Code.ensure_loaded?(api_module)
 
-    if start? && !api_module_present? do
+    if !api_module_present? do
       Logger.error(
         "ExchangeRates api module #{api_module_name(api_module)} could not be loaded. " <>
           "  Does it exist?"
@@ -36,7 +33,7 @@ defmodule Money.Application do
       Logger.warn("ExchangeRates service will not be started.")
     end
 
-    start? && api_module_present?
+    api_module_present?
   end
 
   defp api_module_name(name) when is_atom(name) do
@@ -51,29 +48,42 @@ defmodule Money.Application do
 
   @doc false
   def maybe_log_deprecation do
-    case Application.fetch_env(:ex_money, :exchange_rate_service) do
-      {:ok, start?} ->
+    case Application.fetch_env(:ex_money, :delay_before_first_retrieval) do
+      {:ok, _} ->
         Logger.warn(
-          "Configuration option :exchange_rate_service is deprecated. " <>
-            "Please use :auto_start_exchange_rate_service instead. " <>
-            "This keyword will be removed in ex_money 2.0"
+          "Configuration option :delay_before_first_retrieval is deprecated. " <>
+          "Please remove it from your configuration."
         )
 
-        Application.put_env(:ex_money, :auto_start_exchange_rate_service, start?)
-        Application.delete_env(:ex_money, :exchange_rate_service)
+        Application.delete_env(:ex_money, :delay_before_first_retrieval)
 
       :error ->
         nil
     end
 
-    case Application.fetch_env(:ex_money, :delay_before_first_retrieval) do
+    case Application.fetch_env(:ex_money, :auto_start_exchange_rate_service) do
       {:ok, _} ->
         Logger.warn(
-          "Configuration option :delay_before_first_retrieval is deprecated. " <>
-            "Please remove it from your configuration."
+          "Configuration option :auto_start_exchange_rate_service is deprecated. " <>
+          "Please remove it from your configuration. " <>
+          "Use the :exchange_rates_retrieve_every option instead to define a retrieval period. "
         )
 
-        Application.delete_env(:ex_money, :delay_before_first_retrieval)
+        Application.delete_env(:ex_money, :auto_start_exchange_rate_service)
+
+      :error ->
+        nil
+    end
+
+    case Application.fetch_env(:ex_money, :exchange_rate_service) do
+      {:ok, _} ->
+        Logger.warn(
+          "Configuration option :exchange_rate_service is deprecated. " <>
+          "Please remove it from your configuration. " <>
+          "Use the :exchange_rates_retrieve_every option instead to define a retrieval period. "
+        )
+
+        Application.delete_env(:ex_money, :exchange_rate_service)
 
       :error ->
         nil