+++ lib/money/exchange_rates/exchange_rate_cache.ex
@@ -6,7 +6,9 @@ defmodule Money.ExchangeRates.Cache do
   def init do
     if :ets.info(@ets_table) == :undefined do
       :ets.new(@ets_table, [
-        :named_table, :public, read_concurrency: true
+        :named_table,
+        :public,
+        read_concurrency: true
       ])
     else
       @ets_table
@@ -64,13 +66,15 @@ defmodule Money.ExchangeRates.Cache do
   def store_latest_rates(rates, retrieved_at) do
     :ets.insert(@ets_table, {:latest_rates, rates})
     :ets.insert(@ets_table, {:last_updated, retrieved_at})
-  rescue ArgumentError ->
-     Logger.error("Failed to store latest rates")
+  rescue
+    ArgumentError ->
+      Logger.error("Failed to store latest rates")
   end
 
   def store_historic_rates(rates, date) do
     :ets.insert(@ets_table, {date, rates})
-  rescue ArgumentError ->
-     Logger.error("Failed to store historic rates for #{inspect date}")
+  rescue
+    ArgumentError ->
+      Logger.error("Failed to store historic rates for #{inspect(date)}")
   end
 end